refactor: move createPcg factory into index
Hoists the createPcg / createPcg32 factory (and the MASK_64 constant
used only here) out of pcg32.ts and into index.ts, alongside the
already-exported createMulberry32 path. pcg32.ts now exposes only the
pcg32 algorithm (pcg32Advance, pcg32Output) and resolveStreamScheme,
matching the structure of mulberry32.ts.
